# applications/counter-go-grpc-nextjs-postgres/services/backend/openapi/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BackendService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/version": {
      "get": {
        "operationId": "BackendService_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    },
    "/visit": {
      "post": {
        "operationId": "BackendService_CreateVisit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateVisitResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    },
    "/visit/statistics": {
      "get": {
        "operationId": "BackendService_GetVisitStatistics",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiGetVisitStatisticsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BackendService"
        ]
      }
    }
  },
  "definitions": {
    "apiCreateVisitResponse": {
      "type": "object",
      "properties": {
        "visitNumber": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiDailyVisitStatistics": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "visits": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiGetVisitStatisticsResponse": {
      "type": "object",
      "properties": {
        "totalVisits": {
          "type": "string",
          "format": "int64"
        },
        "visits": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/apiDailyVisitStatistics"
          }
        }
      }
    },
    "apiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "hello": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}